requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.15.0
jpholiday>=1.0.0
PyYAML>=6.0
